+++ src/components.d.ts
@@ -572,7 +572,6 @@ export namespace Components {
     'restFetch'?: RestFetch;
   }
   interface ManifoldServiceCardView {
-    'borderDisabled'?: boolean;
     'description'?: string;
     'isFeatured'?: boolean;
     'isFree'?: boolean;
@@ -1629,7 +1628,6 @@ declare namespace LocalJSX {
     'restFetch'?: RestFetch;
   }
   interface ManifoldServiceCardView extends JSXBase.HTMLAttributes<HTMLManifoldServiceCardViewElement> {
-    'borderDisabled'?: boolean;
     'description'?: string;
     'isFeatured'?: boolean;
     'isFree'?: boolean;